smooth bar chart animation

+++ project/src/application/ITransforms.java
@@ -22,7 +22,7 @@ public interface ITransforms {
 	 * @param p_image
 	 * @return
 	 */
-	public int[] createHistogram(Image p_image);
+	public double[] createHistogram(Image p_image);
 	public WritableImage enableMaskFilter(Image p_image, double[][] kernel);
 	public WritableImage negativeFilter(Image p_image);
 	public WritableImage changeContrast(Image p_image);
